Fix incremental compilation of configuration properties metadata
