+++ inject-java/src/main/java/io/micronaut/annotation/processing/BeanDefinitionInjectProcessor.java
@@ -326,8 +326,6 @@ public class BeanDefinitionInjectProcessor extends AbstractInjectAnnotationProce
                         error("Unexpected error: %s", message != null ? message : e.getClass().getSimpleName());
                     }
                 }
-
-                writeConfigurationMetadata();
             } finally {
                 AnnotationUtils.invalidateCache();
                 AbstractAnnotationMetadataBuilder.clearMutated();
@@ -347,35 +345,6 @@ public class BeanDefinitionInjectProcessor extends AbstractInjectAnnotationProce
         } catch (Exception e) {
             String message = e.getMessage();
             error("Error occurred writing META-INF files: %s", message != null ? message : e);
-        }
-    }
-
-    private void writeConfigurationMetadata() {
-        ConfigurationMetadataBuilder.getConfigurationMetadataBuilder().ifPresent(builder -> {
-            try {
-                if (builder.hasMetadata()) {
-                    ServiceLoader<ConfigurationMetadataWriter> writers = ServiceLoader.load(ConfigurationMetadataWriter.class, getClass().getClassLoader());
-
-                    try {
-                        for (ConfigurationMetadataWriter writer : writers) {
-                            writeConfigurationMetadata(builder, writer);
-                        }
-                    } catch (ServiceConfigurationError e) {
-                        warning("Unable to load ConfigurationMetadataWriter due to : %s", e.getMessage());
-                    }
-                }
-            } finally {
-                ConfigurationMetadataBuilder.setConfigurationMetadataBuilder(null);
-            }
-        });
-
-    }
-
-    private void writeConfigurationMetadata(ConfigurationMetadataBuilder<?> metadataBuilder, ConfigurationMetadataWriter writer) {
-        try {
-            writer.write(metadataBuilder, classWriterOutputVisitor);
-        } catch (IOException e) {
-            warning("Error occurred writing configuration metadata: %s", e.getMessage());
         }
     }
