+++ inject-java/src/main/resources/META-INF/gradle/incremental.annotation.processors
@@ -2,3 +2,4 @@ io.micronaut.annotation.processing.TypeElementVisitorProcessor,dynamic
 io.micronaut.annotation.processing.AggregatingTypeElementVisitorProcessor,dynamic
 io.micronaut.annotation.processing.BeanDefinitionInjectProcessor,dynamic
 io.micronaut.annotation.processing.PackageConfigurationInjectProcessor,dynamic
+io.micronaut.annotation.processing.ConfigurationMetadataProcessor,dynamic
