+++ inject-java/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -2,3 +2,4 @@ io.micronaut.annotation.processing.TypeElementVisitorProcessor
 io.micronaut.annotation.processing.AggregatingTypeElementVisitorProcessor
 io.micronaut.annotation.processing.PackageConfigurationInjectProcessor
 io.micronaut.annotation.processing.BeanDefinitionInjectProcessor
+io.micronaut.annotation.processing.ConfigurationMetadataProcessor
